Move the refparser out to the app level
The schema editor needs it too, probably. The schema editor is
thoroughly broken by these changes, as the schemas from the ConfigDB all
have urn:uuid: IDs. Fixing the editor to support that will cause it to
generate schemas that won't work with ACS installations which still
fetch from Github. We need to choose a way forward here.
