+++ acs-manager/resources/js/components/Schemas/SchemaBrowserOverlay.vue
@@ -7,7 +7,6 @@
   <overlay :show="show" @close="$emit('close')" title="Select Device Schema">
     <template #content>
       <schema-browser :show-new="showNew" v-if="show" :device-schemas="deviceSchemas"
-                      :device-schema-versions="deviceSchemaVersions"
                       @selected="schemaVersionSelected" @new-schema="newSchema"/>
     </template>
   </overlay>
@@ -25,7 +24,6 @@ export default {
   props: {
     show: { required: true, type: Boolean },
     deviceSchemas: { required: true },
-    deviceSchemaVersions: { required: true },
     showNew: { required: false, type: Boolean, default: false },
   },
 
@@ -53,17 +51,17 @@ export default {
       })
     },
 
+    /* XXX bmz: We should not be doing the fetches here. We have two
+     * callers, the origin map editor which wants a deref'd schema and
+     * the schema editor which wants a raw schema. We never need both,
+     * so we should just return the info and leave the fetch to the
+     * caller. */
     async schemaVersionSelected (schemaObj) {
       // Get the raw JSON in case our caller needs that (e.g. Schema Editor)
       const uuid = schemaObj.schema_uuid;
       const raw = await this.refParser.fetchSchema(uuid)
-        .catch(error => {
-          if (error && error.response && error.response.status === 401) {
-            this.goto_url('/login');
-          }
-          this.handleError(error);
-        });
-      const parsed = await this.refParser(`urn:uuid:${uuid}`)
+        .catch(error => this.handleError(error));
+      const parsed = await this.refParser.dereference(uuid)
         .catch(err => this.handleError(err));
 
       this.$emit('schema-selected', {
